Added the ability to send back to back output commands.
While investigating why Tyrian could not use the sound card in SB16 mode
it was observed that the game sent back to back output commands (0x14).
This is simliar to the Monkey Island investigation, but in that case
sending back to back output commands would cause the games internal
counters to be out of sync and hang. Monkey Island expected all bytes to
be transfered before playing the next sound.

In Tyrian it appears sounds stop working if all the bytes are played.
When run in SoundBlaster Pro mode (sbctl.exe t4) the game behaves like
Monkey Island and sends pause commands before interrupting sound with
new output. In SB16 Mode (sbctl.exe t6) it doesn't bother with the pause
commands.

Its possible the game is attempting to avoid an interrupt, always
starting a new sound near the end of the DMA transfer.

Updating the core to internally "pause" the sound so it can be restarted
fixes Tyrian in SB16 mode. Without reference hardware its not clear how
to confirm this is the right change.